requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
openai>=1.68.2,<2.0.0
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.1
pydub==0.25.1
SpeechRecognition==3.10.0
SQLAlchemy>=2.0.25
pytest==7.4.3
pytest-asyncio==0.21.1
loguru==0.7.2
langchain-core>=0.3.58,<1.0.0
langchain-openai>=0.3.16
langchain-text-splitters>=0.3.8
langchain-community>=0.0.20
langsmith>=0.1.125,<0.4 